requirements.txt
asgiref==3.8.1
Django==4.1.13
djangorestframework==3.15.2
djongo==1.3.6
feedparser==6.0.11
pymongo==3.11.4
pytz==2024.2
sgmllib3k==1.0.0
sqlparse==0.2.4
